V1

+++ app/Http/Controllers/AlbumController.php
@@ -24,31 +24,34 @@ class AlbumController extends Controller
 
     public function store(Request $request)
     {
+        $countAlbums = Album::where('id_usuario', Auth::user()->id_usuario)->count();
         if ($request->ajax()) {
-            $album = new Album();
-            $album->nombre = $request->nombre;
-            $album->slug_album = str_slug( $request->nombre . '-' . time() );
-            $album->id_usuario = auth()->user()->id_usuario;
-            $album->save();
-            $folder = "/public/galeria/{$album->slug_album}";
-            Storage::makeDirectory("public/galeria/{$album->slug_album}");
-            return response()->json([
-                "message" => "Album {$album->nombre} creado!",
-                "album" => $album
-            ], Response::HTTP_CREATED);
+            if ($countAlbums <= 5) {
+                $album = new Album();
+                $album->nombre = $request->nombre;
+                $album->slug_album = str_slug( $request->nombre . '-' . time() );
+                $album->id_usuario = auth()->user()->id_usuario;
+                $album->save();
+                $folder = "/public/galeria/{$album->slug_album}";
+                Storage::makeDirectory("public/galeria/{$album->slug_album}");
+                return response()->json([
+                    "message" => "Album {$album->nombre} creado!",
+                    "album" => $album
+                ], Response::HTTP_CREATED);
+            } else {
+                return response()->json(["message" => "!Los siento solo puede tener 6 albums como máximo!"], Response::HTTP_OK);
+            }
         }
     }
 
     public function show(Album $album) {
-        if (auth()->user()->id_usuario == $album->id_usuario) {
-            $imagenes = $album->imagenes;
-            return view('multimedia.album.show', compact('album', 'imagenes'));
-        } else {
-            return view('errors.404');
-        }
+        $this->authorize('view', $album);
+        $imagenes = $album->imagenes;
+        return view('multimedia.album.show', compact('album', 'imagenes'));
     }
 
     public function destroy(Album $album) {
+        $this->authorize('delete', $album);
         $tmp = $album->nombre;
         $album->imagenes()->forceDelete();
         $album->delete();
@@ -59,8 +62,9 @@ class AlbumController extends Controller
     }
 
     public function update(Request $request, Album $album) {
+        $this->authorize('update', $album);
         if ($request->ajax()) {
-            $album->fill($request->all());
+            $album->nombre = $request->name;
             $album->save();
             return response()->json("¡Album actualizado!", Response::HTTP_OK);
         }
